Fix default tag edit/remove button visibility bug

+++ client/src/main/java/client/scenes/ExpenseScreenCtrl.java
@@ -82,6 +82,7 @@ public class ExpenseScreenCtrl implements Initializable, SimpleRefreshable {
     private List<CheckBox> participantCheckBoxes;
     private final ImageUtils imageUtils;
     private final Styling styling;
+    private int i = 0;
 
     /**
      *
@@ -136,55 +137,60 @@ public class ExpenseScreenCtrl implements Initializable, SimpleRefreshable {
                 deleteButton = new Button("", imageUtils.generateImageView("x_remove.png", 15));
                 styling.applyStyling(editButton, "positiveButton");
                 buttonsBox = new HBox(editButton, deleteButton);
-                pane = new AnchorPane(label);
+                pane = new AnchorPane(label, buttonsBox);
                 AnchorPane.setRightAnchor(buttonsBox, 0.0);
             }
 
             @Override
             protected void updateItem(Tag item, boolean empty) {
                 super.updateItem(item, empty);
+                int j = i;
                 if (empty || item == null) {
                     setText(null);
                     setGraphic(null);
                 } else {
                     label.setText(item.getTagName());
+                    System.out.println(j + ": Label of: " + label.getText());
                     label.setStyle("-fx-background-color: " + item.getColorCode() + ";" +
                             "-fx-background-radius: 15;" +
                             "-fx-padding: 5 10 5 10;" +
                             "-fx-text-fill: white;");
-                    if(!item.equals(findDefaultTag(tags))) {
+                    if(findDefaultTag(tags) != null && !item.equals(findDefaultTag(tags))) {
                         editButton.setOnMousePressed(event -> mainCtrl.switchToEditTagScreen(item, expenseId));
                         deleteButton.setOnMousePressed(event -> {
+                            System.out.println(j + ": Pressed delete on: " + item.getTagName());
                             server.deleteTag(currentEvent.getId(), String.valueOf(item.getId()));
                             tagComboBox.getItems().remove(item);
                             tagComboBox.setValue(findDefaultTag(tags));
                         });
-                        if(!pane.getChildren().contains(buttonsBox))
-                            pane.getChildren().add(buttonsBox);
-                    }
+                        buttonsBox.setVisible(true);
+                    } else
+                        buttonsBox.setVisible(false);
                     setGraphic(pane);
                 }
+                System.out.println("----------------");
+                i++;
             }
         });
-        tagComboBox.setButtonCell(new ListCell<>() {
-            @Override
-            protected void updateItem(Tag item, boolean empty) {
-                super.updateItem(item, empty);
-                if (empty || item == null) {
-                    setText(null);
-                    setGraphic(null);
-                } else {
-                    Label label = new Label(item.getTagName());
-                    label.setAlignment(Pos.CENTER);
-                    label.setStyle("-fx-background-color: " + item.getColorCode() + ";" +
-                            "-fx-background-radius: 15;" +
-                            "-fx-padding: 5 10 5 10;" +
-                            "-fx-text-fill: white;");
-
-                    setGraphic(label);
-                }
-            }
-        });
+//        tagComboBox.setButtonCell(new ListCell<>() {
+//            @Override
+//            protected void updateItem(Tag item, boolean empty) {
+//                super.updateItem(item, empty);
+//                if (empty || item == null) {
+//                    setText(null);
+//                    setGraphic(null);
+//                } else {
+//                    Label label = new Label(item.getTagName());
+//                    label.setAlignment(Pos.CENTER);
+//                    label.setStyle("-fx-background-color: " + item.getColorCode() + ";" +
+//                            "-fx-background-radius: 15;" +
+//                            "-fx-padding: 5 10 5 10;" +
+//                            "-fx-text-fill: white;");
+//
+//                    setGraphic(label);
+//                }
+//            }
+//        });
 
         Tag defaultTag = findDefaultTag(tags);
         if (!tags.isEmpty()) {
